refactor!: DuplexStreamChannel is now DuplexStream<T>
DuplexStream<T> directly implements Stream<T> and StreamSink<T>.

This avoids an issue with the current approach where calling
channel.sink.close() would not behave the same as channel.close().
There are other ways to mitigate this, but it just feels like we
are spreading closely related functionality over multiple classes.

This is one of the few situations where I have opted for inheritance
over composition.