Update the copyright year to span 2025. Attribute Dart project authors up until 2024.
Relic is a fork of [shelf](https://github.com/dart-lang/shelf).
However for reasons that are not clear in hindsight, it didn't start as a proper git fork.

Instead it started as new repo (commit 36b87cc4), and all the existing shelf code was copied in
(in commit 76562673).

However there where already changes done, files moved, etc. And - importantly for this commit -
all the copyright notices were removed from the top of the files, except for the LICENSE file.

This is still the case today, but at least this commit attribute copyright up until 2024 to the
Dart project authors.